Entry関連のクラスをまとめた
